app/requirements.txt
streamlit>=1.30.0
pyyaml>=6.0
card-generator
